new Version of Networkparser and fix refactoring

+++ src/org/sdmlib/models/debug/FlipBook.java
@@ -265,7 +265,7 @@ public class FlipBook implements MapUpdateListener,  PropertyChangeInterface
          
          JsonObject jsonObject = new JsonObject().withValue(line);
          
-         long value = jsonObject.getLong("stopStep");
+         long value = jsonObject.getInt("stopStep");
          
          stopStep = value;
